Update app.rb file
