--- src/data/author_data.rb
@@ -1,36 +1,0 @@
-require 'json'
-require_relative '../classes/author'
-
-module AuthorsData
-  def read_authors
-    data = []
-    file = './src/storage/authors.json'
-    if File.exist?(file)
-      JSON.parse(File.read(file)).each do |author|
-        data.push(Author.new(author[ 'first_name' ], author[ 'last_name' ]))
-      end
-    else
-      File.write(file, [])
-    end
-
-    data
-  end
-
-  def save_author
-    data = []
-    @music_albums.each do |author|
-      data.push({ first_name: author.first_name, last_name: author.last_name })
-    end
-    open('./src/storage/authors.json', 'w') { |f| f << JSON.generate(data) }
-  end
-
-  def create_author
-    puts"\nEnter the last name:"
-    lname= gets.chomp
-    puts"\nEnter the first name:"
-    fname= gets.chomp
-    data.push(Author.new(lname, fname))
-
-    puts"\nAuthor created successfully."
-  end
-end
