+++ src/modules/author_data.rb
@@ -1,7 +1,7 @@
 module AuthorsData
   def read_authors
     data = []
-    file = './data/authors.json'
+    file = './src/storage/authors.json'
     if File.exist?(file)
       JSON.parse(File.read(file)).each do |author|
         data.push(Author.new(author['first_name'], author['last_name']))
@@ -18,6 +18,16 @@ module AuthorsData
     @music_albums.each do |author|
       data.push({ first_name: author.first_name, last_name: author.last_name })
     end
-    open('./data/authors.json', 'w') { |f| f << JSON.generate(data) }
+    open('./src/storage/authors.json', 'w') { |f| f << JSON.generate(data) }
+  end
+
+  def create_author
+    puts"\nEnter the last name:"
+    lname= gets.chomp
+    puts"\nEnter the first name:"
+    fname= gets.chomp
+    data.push(Author.new(lname, fname))
+
+    puts"\nAuthor created successfully."
   end
 end
